revert separate strategies: engines will extend the caffe standards
